refactor: convert profile page to server component ♻️
- Split client components into modular components in _components
- Create PersonalInfoForm, SecurityForm, AvatarSection components
- Convert main page to server component with proper auth handling
- Add metadata for SEO optimization
- Improve code organization and maintainability
